refactor: follow ddd structure
