+++ src/lib/core/application/service/store/global.dto.ts
