+++ src/lib/core/application/service/store/global.store.ts
