+++ src/lib/core/domain/models/social.dto.ts
