+++ src/lib/core/infrastructures/repository/social.repository.ts
@@ -1,11 +1,11 @@
 import type { AxiosInstance, AxiosResponse, AxiosError } from 'axios'
-import type { ISocialRepository } from '$lib/domain/repository/social.repository'
+import type { ISocialRepository } from '$lib/core/domain/repository/social.repository'
 import type {
 	CreatePostRequest,
 	CreatePostResponse,
 	CommentResponse,
 	PostResponse
-} from '$lib/domain/models/social.dto'
+} from '$lib/core/domain/models/social.dto'
 
 export class SocialRepository implements ISocialRepository {
 	private readonly $axios: AxiosInstance
