+++ src/lib/core/interfaces/components/Loading.svelte
