+++ src/lib/core/interfaces/components/SEO.svelte
