+++ src/routes/+page.ts
@@ -1,4 +1,4 @@
-import { socialAPI } from '$lib/application/service/store/social.api.store'
+import { socialAPI } from '$lib/core/application/service/store/social.api.store'
 import type { PageLoad } from './$types'
 
 export const load: PageLoad = async () => {
