+++ src/routes/post/+page.ts
@@ -1,5 +1,5 @@
-import { socialAPI } from '$lib/application/service/store/social.api.store'
-import { homeStore } from '$lib/module/home/application/service/store/home.store'
+import { socialAPI } from '$lib/core/application/service/store/social.api.store'
+import { homeStore } from '$lib/core/module/home/application/service/store/home.store'
 import type { PageLoad } from './$types'
 
 export const load: PageLoad = async () => {
